Log akka errors
