+++ neo-cli/neo-cli.csproj
@@ -28,7 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Neo" Version="3.0.0-CI00892" />
+    <PackageReference Include="Neo" Version="3.0.0-CI00904" />
   </ItemGroup>
 
   <ItemGroup>
